timekeeper: Add basic jira interaction
This commit introduces a new tab in the timekeeper TUI which is still
under construction. Currently it is possible to select a date, for which
the current JIRA sprint is requested and all issues in the sprint(s) of
the given date are retrieved. In the next step all timeflip entries'
descriptions are parsed for JIRA-IDs and if such are found, a booking
suggestion for the issue is created. If multiple JIRA-IDs are part of a
description the duration of the timeflip entry is split-up into multiple
booking suggestions.

A bunch of new configuration options are added to provide the user to
specify the correct JIRA connection.

Now it is also possible to configure a setting in the .toml file named
[[sides_jira_matching]] to be able to book timeflip entries like this:

(Side1-Conference) 10.11 00:10:00 08:00-08:10 [No-Issue-ID in desc.]

We might want to book all entries of a timeflip-side to a specific issue
in the current JIRA sprint. We therefore can configure a timeflip-side
to look for an issue in the sprint with a specific summary. If the
configured regular expression matches the summary of an issue this
JIRA-ID is used for the booking suggestion.

With the current implementation state the booking suggestions are only
displayed but not actually booked. This is subject to a future commit.