timekeeper: Load issue regex from config
With this commit the regex which is used to identify Jira's "issue
numbers" in the descriptions is no longer hardcoded. It is now possible
to configure the "booker_issue_regex=" setting in timeflip.toml. Be
adviced it might be necessary to use single quotes as otherwise the toml
loader might have problems with the escaping.